Split song context menu views
